events: moved message handler to event handlers
