+++ Roki.Core/Services/Roki.cs
@@ -100,7 +100,7 @@ namespace Roki
             service.AddHttpClient();
 
             service.LoadFrom(Assembly.GetAssembly(typeof(CommandHandler)));
-            service.LoadFrom(Assembly.GetAssembly(typeof(MessageService)));
+            service.LoadFrom(Assembly.GetAssembly(typeof(EventHandlers)));
 
             Services = service.BuildServiceProvider();
             var commandHandler = Services.GetService<CommandHandler>();
@@ -181,10 +181,10 @@ namespace Roki
             stats.Initialize();
             var commandHandler = Services.GetService<CommandHandler>();
             var commandService = Services.GetService<CommandService>();
-            var messageLogger = Services.GetService<MessageService>();
+            var eventHandlers = Services.GetService<EventHandlers>();
 
             await commandHandler.StartHandling().ConfigureAwait(false);
-            await messageLogger.StartService().ConfigureAwait(false);
+            await eventHandlers.HandleEvents().ConfigureAwait(false);
 
             var _ = await commandService.AddModulesAsync(GetType().GetTypeInfo().Assembly, Services).ConfigureAwait(false);
